Huge update.
Removed Card class. All Trello objects will now use the TrelloObj class.

Trello objects return Proxies with property accessors overwritten. Property access now always returns a Promise. When accessing a property, if that property hasn't yet been retrieved from Trello, it will happen asynchronously and return a Promise. Data that has already been downloaded will be returned as a resolved Promise. Properties that aren't Trello properties will never request data from Trello, and will instead return a resolved Promise of whatever had already been set (or will resolve to undefined if it's new).

There is a map of property names to instructions for each Trello type. So far, I've only created Card and Action, and those are incomplete. Eventually there will be one for every Trello object type. These will have to be maintained manually.

These maps have two types of properties. Those that can be retrieved simply by accessing the object at the given ID, and those that need additional requests to Trello to be accessed. This is all seamless to the user.

As a consequence to all of this, node now must be started with the --harmony --harmony-proxies command line options. The reason for the latter should be obvious. For the former, it's because I'm using the string.endsWith() function.

The next step is to flesh out these maps and test that the various kinds of access are all working.

After that, I will work on setting data and saving it back to Trello.

Once core functionality is complete, I will build a version that can run on the current version of node without the additional command line options. This will uglify the interface since I won't be able to modify property getters and setters using Proxies. Though what I'm doing with that (returning all properties as Promises) is probably anathema anyway. Though, if I take long enough, node may get upgraded to include harmony functionality by default and I won't have to worry about making this library backwards compatible.