Implemented card class
The card class is essentially complete, though there's a few more convenience
methods I could (and probably will) add. Everything seems pretty awesome and
this project is looking like it'll be highly useful.

With this commit I also created a factory which will handle all trello object
creation. It was added to aid with integrating the card class, and adding
more Trello objects will be a cinch.