Implemented more stuff for card
Card is almost fully complete, including documentation. All that's left is to
add a convenience creator function. I also implemented action, which is much
simpler, and that's 100% complete.

I added a test to make sure that constructor functions were passed in to
create Trello objects, otherwise an error will be thrown.