Implemented webhook
I also made a getter for each object type on trello.js and made the
general purpose get private. This will reduce user errors and makes
the interface easier to use. In doing this I had to remove all tests
from trello, which is fine, as there's almost no real logic in it.

All coding is complete for what should be version 1 of the software.
What's left is building the Readme.md, generating the documentation,
transpiling the code, packaging everything in npm, and publishing.
Exciting!