Implemented the ability to skip local validation
get(), set(), and del() on trelloObj all now accept an optional
skipValidation property. This allows the user to skip any local
validation and access the API with any path. This gives our
library 100% coverage of the Trello API, and means that any
Trello updates are immediately picked up without any effort on our
part. Granted, convenience functions will be lacking. We're still
missing validation for GETs that aren't a simple property, though.

I also worked on the README.md, which is now nearly complete. I
still need to generate documentation and link to it from the
README